# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APP-STOCK-LNB API",
        "description": "Description de votre API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/intervention-vehicules": {
            "get": {
                "tags": [
                    "Intervention Véhicule"
                ],
                "summary": "Lister les interventions des véhicules",
                "operationId": "aed7252502a5fb1e9a793bb3b0cdd360",
                "responses": {
                    "200": {
                        "description": "Liste récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des interventions de véhicules"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/InterventionVehicule"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Intervention Véhicule"
                ],
                "summary": "Créer une nouvelle intervention de véhicule",
                "operationId": "ad72552442adeb4acd55e0817a815c31",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicule_id",
                                    "titre",
                                    "montant",
                                    "date_intervention",
                                    "type_intervention_id"
                                ],
                                "properties": {
                                    "vehicule_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "titre": {
                                        "type": "string",
                                        "example": "Vidange"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25000
                                    },
                                    "observation": {
                                        "type": "string",
                                        "example": "RAS"
                                    },
                                    "date_intervention": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-15"
                                    },
                                    "type_intervention_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "date_expiration": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Intervention créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResourceInterventionResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/intervention-vehicules/{id}": {
            "put": {
                "tags": [
                    "Intervention Véhicule"
                ],
                "summary": "Mettre à jour une intervention",
                "operationId": "94b477d302da30d30302c623cba73b2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'intervention",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InterventionVehicule"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Intervention mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResourceInterventionResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Intervention Véhicule"
                ],
                "summary": "Supprimer une intervention",
                "operationId": "ca99845c48cab21ff450f06d734e1ba4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Intervention supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "intervention supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InterventionVehicule": {
                "title": "InterventionVehicule",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "vehicule_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "titre": {
                        "type": "string",
                        "example": "Réparation moteur"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 85000
                    },
                    "observation": {
                        "type": "string",
                        "example": "Pièce changée"
                    },
                    "date_intervention": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-07-01"
                    },
                    "date_expiration": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-12-01"
                    },
                    "type_intervention_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PostResourceInterventionResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Intervention créée avec succès"
                    },
                    "data": {
                        "$ref": "#/components/schemas/InterventionVehicule"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}